Tentative fixes and a new approach to the generator.
Tried fixing the dungeon generation and actually
removed quite a few bugs - at least it doesn't crash nor gets stuck on a loop. But it still doesn't work.

Since it seems like too much time would be spent trying to fix the current generator, implementation of a new and simpler approach is on the way.