Prototype
Edited the dungeon drawer script to better handle the current algorithm and started making a simple prototype in order to test the actual game mechanics. For now, a character (from the asset store) is able to move around the dungeon, with a camera following it and full collision with walls and floor.