Movement and Animation!
Improved dungeon drawing (using quads instead of cubes for covering up inaccessible places).

Added a placeholder model from the asset store, implemented basic controls for it and added an outline shader so the character can be seen through walls.

Made the camera and character movement much smoother and added a very simple (not yet functional) HUD.

A lot of work was done "under the hood" as preparation for the upcoming combat system (that will hopefully be implemented in the near future), mainly having to do with character attributes and their inner workings (going for a modular-ish system that is as flexible as I want it to be).