Custom Editor and Interactions
Created a script to draw character properties in a nicer way. Can use a lot of improvements, but it gets the job done and is not the main concern of the project (also, a lot of unnecessary hours were already dumped into this because of reasons).

Made some changes to the inner workings of characters and added another model (from the Asset Store) to act as an enemy. Currently, said enemy just stares at the player in a very creepy fashion and may be kicked around the map.