COMBAT! (finally some action up in here...)
Added a "hit" animation to the creepy enemy so it would react a little more naturally to being kicked around and a "death" animation too.

Together with the death animation, implemented a basic healthbar system for the enemy (using a world space canvas with some nifty animation) and a simple combat system (get the player damage and deduce it from the enemy health).

Improved DungeonDrawer (renamed to SimpleDungeonDrawer) and started adding custom models to it (for now, just a spiky doorway, which so happens to be the first original asset being used - hurray!).