Maintenance
Fixed a bug where, when  it was to be trimmed, the end coordinates of a corridor would be used as its size, resulting in large chunks of the dungeon being replaced by walls.

Fixed a bug in the "GetRoomEdgeCell" method where it would get the wrong edge.

Updated variable names to follow a convention (underscore before private variables)

Added different debugging levels to more easily adjust the kind of messages that are shown in the console.

Added start and exit points for the dungeons (the exit is rendered as a hole with a trigger, but currently does nothing).