Level Progression
Made it so a new dungeon is generated once the player hits the exit (with a cool little falling transition).

Added a "falling" state to the character mecanim controller.

Vertical speed is now ignored when deciding if the walking animation should be played.

CharacterController now verifies if the character is on solid ground and, if not, locks its movement.